Rename widgetbook_golden_alchemist_renderer.dart file
